+++ src/MiniProfiler.Shared/Internal/Render.cs
@@ -56,10 +56,12 @@ namespace StackExchange.Profiling.Internal
 
             sb.Append("\" data-position=\"");
             sb.Append((renderOptions?.Position ?? options.PopupRenderPosition).ToString());
-            sb.Append('"');
 
             sb.Append("\" data-scheme=\"");
             sb.Append((renderOptions?.ColorScheme ?? options.ColorScheme).ToString());
+
+            sb.Append("\" data-decimal-places=\"");
+            sb.Append((renderOptions?.DecimalPlaces ?? options.PopupDecimalPlaces).ToString());
             sb.Append('"');
 
             if (isAuthorized)